[STD] 'make' should default to 'make all'
While make itself doesn't enforce this, the history of make culture
is that people who want to make subsets should bear the burden of
specifying an additional argument to make.  It's not like it's that
hard to run the previous make command.  And in this case, the STD_P5
regexes are part of the official definition of STD, even if the
rest of the P5 file is still 90% crud.  :)