recover from several months of specrot
The bootstrap of STD was completely hosed by the sigspace and +%
changes, as well as various other dysfunctions.  We now have a complete
bootstrap of the new sigspace model, and a partial implementation of +%
(sufficient only to bootstrap a STD containing +% rather than **).

With the new sigspace rules we can rewrite many of the tokens to rules
and still preserve LTM correctness.