clean up misleading P5 special var warnings
Remove the tests for rare variables, especially those that will become
syntax errors anyway, like $) and friends.  This also allows us to parse
bare sigils as rvalues, so we don't have to distinguish ($a,*,$c) = ...
from my ($a,$,$c) = ... anymore.  That is, if you use a bare $ as an
rvalue, we'll simply declare an anonymous variable for you now.  If we
make the declarator default to 'state', it might even be useful as $++.